client.py: fix power checks on desktop systems
The path /sys/class/power_supply/AC doesn't exist on desktop systems (which
necessarily are on AC power). If the client configuration specifies that AC
power is required, this check would throw an exception on desktop systems
while attempting to read a nonexistent path. Treat the lack of a power_supply
path as "on AC power" to handle this case.